# Optimize partition_validity function used in sort kernels