@zhuqi-lucas 2025-07-16
Pretty good @jhorstmann , i tested locally for one case, it is actually 1.2 faster:

```rust
sort i32 nulls to indices 2^10
                        time:   [3.9156 µs 3.9226 µs 3.9306 µs]
                        change: [−21.573% −21.387% −21.194%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe
```