@jhorstmann 2025-07-16
> The bechmak results also look good to me -- the only one that reports something slow is already so fast I think it is mostly measurement error
> 
> > sort primitive run 2^12                                 1.00      6.4±0.02µs        ? ?/sec    1.12      7.2±0.02µs        ? ?/sec

I agree, probably measurement overhead. But `sort_to_indices` for run end encoded arrays does actually call `partition_validity` and then ignores the result, so it's also possible that the compiler previously optimized that away. I'll take another look and maybe handle ree arrays earlier in that function.